+++ ChangeLog.md
@@ -20,6 +20,10 @@ and this project adheres to
 - `doc` function to use *Levenshtein distance* if the exact symbol is not found 
   to help with typos.  
 
+### Fixed
+
+- a namespace problem with protocols
+
 
 
 ## [1.10.0] - 2021-10-01